# api/api.yaml
swagger: '2.0'
info:
  description: REST API Issue Tracker documentation - Grup 13 D
  version: '1.0'
  title: REST API Issue Tracker
  contact:
    name: Sourcer(Bitbucket)
    url: 'https://bitbucket.org/ademait/aswg13drest/src/master/'
host: grup13aswd-rest-api-blabla.herokuapp.com
basePath: /
...
paths:
...
  /api/v1/users:
    get:
      tags:
        - User
      summary: Busca tots els usuaris
      description: Busca tots els usuaris creats fins al moment
      operationId: getAllUsersUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT: []
      deprecated: false
    post:
      tags:
        - User
      summary: Crea un usuari
      description: crea un usuari amb les dades especificades
      operationId: addUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: userDTO
          description: userDTO
          required: true
          schema:
            $ref: '#/definitions/UserDTO'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT: []
      deprecated: false
  /api/v1/users/apikey:
    get:
      tags:
        - User
      summary: Demana la Api-Key del usuari amb aquest username
      description: retorna la api-key del usuari amb aquest username
      operationId: getUserApiKeyUsingGET
      produces:
        - '*/*'
      parameters:
        - name: username
          in: query
          description: username
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT: []
      deprecated: false
  /api/v1/users/username:
    get:
      tags:
        - User
      summary: Busca un usuari pel username
      description: Busca l'usuari amb el username especificat
      operationId: getUserByUsernameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: username
          in: query
          description: username
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT: []
      deprecated: false
  '/api/v1/users/{id}':
    get:
      tags:
        - User
      summary: Busca un usuari per id
      description: Busca l'usuari amb el id especificat
      operationId: getUserByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT: []
      deprecated: false
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  User:
    type: object
    properties:
      apikey:
        type: string
      id:
        type: integer
        format: int64
      username:
        type: string
    title: User
  UserDTO:
    type: object
    properties:
      apikey:
        type: string
      username:
        type: string
    title: UserDTO